add dlls using contextmanager
